refactor(ranking-ui): move components to a separate directory
