+++ lib/views/pages/home/content/ranking/components/ranking_card.dart
